hide discard for nonauthors
